Bypass GitHub's image cache:
The problem appears to be that GitHub replaces all external images with a cache
stored on their servers. When using the Markdown API, these image URLs are still
replaced, but image doesn't always resolve. The effect is that sometimes images
are shown, and sometimes not. This is probably why #50 appeared to be resolved.

This fix makes sure the image is shown from the source, so it'll always
resolve (as long as it still exists on the original servers). A slightly better
fix would be to only handle images that error out, but that optimization will
be in a future fix.