Clarify that this sends your markdown to GitHub
The current wording allows for the misinterpretation that the rendering is being done locally.
"Github markdown API" could mean a library published by GitHub, and "The styles come directly from GitHub" could mean just the CSS stylesheet.
This should be corrected, because if misinterpreted this way it leads to 2 false assumptions:
- Your data is not being sent to a third party
- This will work offline

I know information later in the readme talks about API rate limits and such, but many people won't read past the Usage section (myself included apparently). I think in particular the repeated mentions of a "local readme file" gave me this mistaken impression.

I think mentioning REST and saying that the rendering is done by GitHub is sufficient, but there may be some other change that clarifies this better.