Fix missing styles by fixing the regular expression to allow adjacent matches
The regular expression worked before because the style elements were separated by a whitespace or newline. This fix works
around the "adjacent matches" issue in `re` by allowing the tag close character, after matching, to be left unconsumed by
the engine. This character is then used to separate the current match from the next, since it may be adjacent, which lets
the engine work as it did before.

Note that it's probably worth using a different parsing approach. Other approaches will be explored in future versions.