Fix scrollToHash failing on links
This change fixes the issue with <a...> anchor tags not scrolling to
their corresponding hash when clicked.

The version of the function `scrollToHash` in templates/index.html is
using `getElementById`, which fails for anchor tags such as `<a
name="foo">`; whereas the homonymously named function in
templates/limit.html (using `getElementsByName` instead) does work
correctly.

This commit merges these two functions with the same name but differing
function body into the correctly working version from
templates/limit.html, and moves it to templates/base.html, such that it
can be shared and extended by the Flask templates accordingly.

Note that as this change modified the templates, the tests must be
regenerated by running the tests/regenerate.py script in order for them
to pass.